compute: move mz_join_core into its own module
